pyproject.toml
[project]
name = "yahoo-mail-sorter"
version = "0.1.0"
description = "CLI tool to classify and sort Yahoo Japan Mail via IMAP"
readme = "README.md"
authors = [
    { name = "rmondo", email = "ryosuke.mondo@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "rich>=14.3.3",
    "typer>=0.24.1",
]

[project.scripts]
yahoo-mail-sorter = "yahoo_mail_sorter.cli:app"

[build-system]
requires = ["uv_build>=0.9.2,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.15.2",
    "types-pyyaml>=6.0.12.20250915",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--strict-markers -v"

[tool.ruff]
target-version = "py313"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "TCH"]

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.coverage.run]
source = ["yahoo_mail_sorter"]

[tool.coverage.report]
show_missing = true
fail_under = 80
